Attempt to optimise a few more things for Firebird

+++ Models/Storage/CleanCommand.cs
@@ -3,6 +3,8 @@ using APSIM.Shared.JobRunning;
 using APSIM.Shared.Utilities;
 using System.Linq;
 using System.Threading;
+using System;
+using System.Globalization;
 
 namespace Models.Storage
 {
@@ -78,30 +80,60 @@ namespace Models.Storage
         /// <param name="currentID">ID of the "Current" checkpoint.</param>
         private void CleanTable(string tableName, string simulationIDs, string simulationNames, int currentID)
         {
-            writer.Connection.BeginTransaction();
-            try
+            var fieldNames = writer.Connection.GetColumnNames(tableName);
+            if (writer.Connection is Firebird)
             {
-                var fieldNames = writer.Connection.GetColumnNames(tableName);
+                // We need to do things a bit differently in Firebird, as it has a limit of 1500 items 
+                // in the WHERE IN() clause, and with things like Wheat Validation we exceed that limit.
+
                 if (fieldNames.Contains("SimulationID") && fieldNames.Contains("CheckpointID"))
                 {
-                    string sql = $"DELETE FROM [{tableName}] " +
-                                 $"WHERE \"SimulationID\" in ({simulationIDs}) ";
-                    if (currentID != -1)
-                        sql += $"AND \"CheckpointID\" = {currentID}";
-                    writer.Connection.ExecuteNonQuery(sql);
-                }
-                else if (fieldNames.Contains("SimulationName") && fieldNames.Contains("CheckpointID"))
-                {
-                    string sql = $"DELETE FROM [{tableName}] " +
-                                 $"WHERE \"SimulationName\" in ({simulationNames}) ";
-                    if (currentID != -1)
-                        sql += $"AND \"CheckpointID\" = {currentID}";
-                    writer.Connection.ExecuteNonQuery(sql);
+                    if (String.IsNullOrEmpty(simulationIDs))
+                        return;
+                    int[] Ids = simulationIDs.Split(',').Select(int.Parse).ToArray();
+                    int start = 0;
+                    int spanLength = 1499;
+                    int arrayLength = Ids.Length;
+                    while (start < arrayLength)
+                    {
+                        int useLength = Math.Min(spanLength, arrayLength - start);
+                        Span<int> span = new Span<int>(Ids, start, useLength);
+                        var simulationIDsCSV = StringUtilities.Build(span.ToArray(), ",");
+                        string sql = $"DELETE FROM [{tableName}] " +
+                                     $"WHERE \"SimulationID\" in ({simulationIDsCSV}) ";
+                        if (currentID != -1)
+                            sql += $"AND \"CheckpointID\" = {currentID}";
+                        writer.Connection.ExecuteNonQuery(sql);
+                        start += useLength;
+                    }
                 }
             }
-            finally
+            else
             {
-                writer.Connection.EndTransaction();
+                writer.Connection.BeginTransaction();
+                try
+                {
+                    if (fieldNames.Contains("SimulationID") && fieldNames.Contains("CheckpointID"))
+                    {
+                        string sql = $"DELETE FROM [{tableName}] " +
+                                     $"WHERE \"SimulationID\" in ({simulationIDs}) ";
+                        if (currentID != -1)
+                            sql += $"AND \"CheckpointID\" = {currentID}";
+                        writer.Connection.ExecuteNonQuery(sql);
+                    }
+                    else if (fieldNames.Contains("SimulationName") && fieldNames.Contains("CheckpointID"))
+                    {
+                        string sql = $"DELETE FROM [{tableName}] " +
+                                     $"WHERE \"SimulationName\" in ({simulationNames}) ";
+                        if (currentID != -1)
+                            sql += $"AND \"CheckpointID\" = {currentID}";
+                        writer.Connection.ExecuteNonQuery(sql);
+                    }
+                }
+                finally
+                {
+                    writer.Connection.EndTransaction();
+                }
             }
         }
     }
